initializeConfig

Checks whether configuration already has been set. If not, it will be initialized with default settings or (if available) system settings. All madrat folders (see setConfig for documentation which folders are available) will be set to the system environment variables MADRAT_SOURCEFOLDER, MADRAT_CACHEFOLDER, etc. if they exist, NA otherwise. NA means subfolders of the mainfolder are used.

Usage

initializeConfig(verbose = TRUE)

Arguments

verbose

boolean deciding whether status information/updates should be shown or not

See also

Author

Jan Philipp Dietrich