May All Data be Reproducible And Transparent (madrat) *

Package provides a basic framework which should improve reproducibility and transparency in data processing. It provides functionality such as automatic meta data creation and management, rudimentary quality management, data caching, work-flow management an data aggregation.

Details

* The title is a wish not a promise. By no means we expect this package to deliver everything what is needed to achieve full reproducibility and transparency, but we believe that it supports efforts in this direction.

See also

Author

Maintainer: Jan Philipp Dietrich dietrich@pik-potsdam.de (ORCID) (Potsdam Institute for Climate Impact Research)

Authors:

Pascal Sauer pascal.sauer@pik-potsdam.de (ORCID) (Potsdam Institute for Climate Impact Research)

Lavinia Baumstark lavinia@pik-potsdam.de (Potsdam Institute for Climate Impact Research)

Stephen Wirth wirth@pik-potsdam.de (Potsdam Institute for Climate Impact Research)

Anastasis Giannousakis giannou@pik-potsdam.de

Renato Rodrigues Renato.Rodrigues@pik-potsdam.de (Potsdam Institute for Climate Impact Research)

Benjamin Leon Bodirsky bodirsky@pik-potsdam.de (Potsdam Institute for Climate Impact Research)

Debbora Leip leip@pik-potsdam.de (Potsdam Institute for Climate Impact Research)

Michael Crawford crawford@pik-potsdam.de (Potsdam Institute for Climate Impact Research)

Ulrich Kreidenweis kreidenweis@pik-potsdam.de

David Klein dklein@pik-potsdam.de (Potsdam Institute for Climate Impact Research)

Patrick Rein patrick.rein@pik-potsdam.de (ORCID) (Potsdam Institute for Climate Impact Research)