Tool: CodeLabels

This function replaces a hash code (e.g. regioncode) or another cryptic code with a human readable code via a given dictionary. This can be useful to make outputs better readable in cases where hash codes are already known to the user. If no entry exists in the dictionary the hash code is returned again.

Usage

toolCodeLabels(get = NULL, add = NULL)

Arguments

get

A vector of hash codes which should be replaced

add

Additional entries that should be added to the dictionary. Need to be provided in the form of a named vector with the structure c(<label>=<hash>), e.g. c(h12="62eff8f7")

Value

A vector with either labels (if available) or hash codes (if no label was available).

See also

Author

Jan Philipp Dietrich

Examples

toolCodeLabels("62eff8f7")
#> [1] "h12"